womens-soccer-readiness/requirements.txt
# Core dependencies
fastapi==0.110.3
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4

# Database
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Data processing
pandas==2.2.2
numpy==1.26.4
scipy==1.14.0

# File handling
python-dateutil==2.9.0
openpyxl==3.1.5

# Development
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0

# Monitoring & Logging
sentry-sdk==2.7.1

# CORS
fastapi-cors==0.0.6